Refactor: move all announcement features to a single package
